Fix PluginManager.validMinimumPluginVersionExists
Re issue #130

The method simply passed parameters in the wrong order. This caused plugins to verify correctly if they had the same version number, but they would be validated if the version number was too low, and rejected if their version number was beyond the requested version. I tidied things up a little while I was at it.