Let MasterClock return the correct time on the first frame
Re issue #158

By leaving `now=0` in the constructor, and `_started=Date.now()`, the MasterClock assumed that the clock started, then immediately jumped back 45 years to 1970 and the dawn of Unix time. This could cause some squirrely results. We've already caught problems in Animation and the particles plugin. I've set `now=Date.now()` in the constructor, which seems to fix everything.