Remove glMatrix calls in GLRenderManager camera code.
The glMatrix proved to be inefficient compared to our implementations. As it's only used in one place in our code, we can probably save a bit of loading time for web apps by removing the glMatrix library altogether. This is the first step: replacing that functionality with native Kiwi methods. Subsequent commits should eliminate the entire glMatrix library.