Make `GLRenderManager.camMatrix` a persistent object.
This turns out to be necessary for some shaders to correctly update the camera matrix if there are no other objects in the scene. The camMatrix is only passed when the shader is enabled, which does not recur when there are no other shaders to switch between. Subsequent draw calls will use the camMatrix object passed from the GLRenderManager. If it creates a new object every frame, that information never makes it to the shader, and the camera does not update. I've updated the camMatrix to be a persistent object, so everything operates as expected.